refactor: useNote 커스텀훅 분리
